address partial feedback
